+++ Pipfile
@@ -27,6 +27,7 @@ PyMySQL = "*"
 Flask-Migrate = "*"
 Unidecode = "*"
 flask-classful = "*"
+jsonify = "*"
 
 [requires]
 python_version = "3.7"
