+++ app/controllers/trial_recipes.py
@@ -1,0 +1,19 @@
+from flask import render_template as template
+from flask_classful import FlaskView
+
+from app.models.users import User
+from app.models.ingredients import Ingredient
+
+
+class TrialRecipesView(FlaskView):
+    def show(self):
+        active_diets = User.load(
+            "ketocalc.jmp@gmail.com", load_type="username"
+        ).active_diets
+        ingredients = Ingredient.load_all_by_author("basic")
+        return template(
+            "recipe/new.html.j2",
+            ingredients=ingredients,
+            diets=active_diets,
+            is_trialrecipe=True,
+        )
