--- app/errors/__init__.py
@@ -1,5 +1,0 @@
-from app.errors.routes import errors_blueprint
-
-
-def create_module(app, **kwargs):
-    app.register_blueprint(errors_blueprint)
