--- app/main/__init__.py
@@ -1,5 +1,0 @@
-from app.main.routes import main_blueprint
-
-
-def create_module(app, **kwargs):
-    app.register_blueprint(main_blueprint)
