--- app/support/__init__.py
@@ -1,5 +1,0 @@
-from app.support.routes import support_blueprint
-
-
-def create_module(app, **kwargs):
-    app.register_blueprint(support_blueprint)
